+++ docs/release-notes.rst
@@ -32,6 +32,7 @@ Features
    - `CPS-638 <https://jira.onap.org/browse/CPS-638>`_  Delete operation for datastore pass through running
    - `CPS-741 <https://jira.onap.org/browse/CPS-741>`_  Re sync after removing cm handles
    - `CPS-780 <https://jira.onap.org/browse/CPS-780>`_  Add examples for parameters, request and response in openapi yaml for cps-core
+   - `CPS-741 <https://jira.onap.org/browse/CPS-777>`_  Ensure all DMI operations use POST method
 
 Bug Fixes
 ---------